Create a standalone withUIState module
